# src/docs/swagger.yaml
openapi: 3.0.3
info:
    title: Computer API
    version: 1.0.0
    description: |
        REST API for discovering and processing events, configuring aggregators, and computing metrics.

        Protected operations accept JWT bearer tokens issued for this API. Depending on the
        operation, the token must represent either a user with the documented minimum system
        role or an authenticated service.
servers:
    - url: /
...
paths:
...
    /api/v1/events:
        get:
            security:
                - UserBearerAuth: []
            tags:
                - Events
            summary: Get all events
            operationId: getEvents
            description: |
                Returns all available event definitions.
                Requires a user token with the `ADMIN` or `SUPERADMIN` system role.
            responses:
                '200':
                    description: Events retrieved successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: '#/components/schemas/Event'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/events/{eventId}:
        parameters:
            - $ref: '#/components/parameters/eventId'
        get:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Events
            summary: Get event by id
            operationId: getEventById
            description: Returns one event definition by ID. Accepts either a user token or a service token.
            responses:
                '200':
                    description: Event retrieved successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/Event'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/events/{eventId}/process:
        parameters:
            - $ref: '#/components/parameters/eventId'
            - $ref: '#/components/parameters/expand'
        post:
            security:
                - UserBearerAuth: []
            tags:
                - Events
            summary: Process an event
            operationId: processEvent
            description: |
                Processes an event for a date and window using the provided fetcher and process
                configurations. Set `expand=true` to include raw fetcher data.
                Requires a user token with the `SUPERADMIN` system role.
            requestBody:
                $ref: '#/components/requestBodies/ProcessEventBody'
            responses:
                '200':
                    description: Event processed successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              oneOf:
                                                  - $ref: '#/components/schemas/ProcessedEventSummary'
                                                  - $ref: '#/components/schemas/ProcessedEventDetailed'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFound'
                '422':
                    $ref: '#/components/responses/ExternalServiceError'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/events/{eventId}/validate:
        parameters:
            - $ref: '#/components/parameters/eventId'
        post:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Events
            summary: Validate event configuration
            operationId: validateEvent
            description: |
                Checks whether the event exists and validates any supplied fetcher and process
                configurations. An unknown `eventId` is reported in the successful response as
                `valid: false`.

                Accepts a service token or a user token with the `SUPERADMIN` system role.
            requestBody:
                $ref: '#/components/requestBodies/ValidateEventBody'
            responses:
                '200':
                    description: Validation executed.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/EventValidationResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '422':
                    $ref: '#/components/responses/ExternalServiceError'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ──────────────────────────────────────────────
    # Aggregators
    # ──────────────────────────────────────────────
...
components:
    securitySchemes:
        UserBearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: |
                JWT whose `type` claim is `user`. Operations also state the minimum accepted
                `systemRole` (`ADMIN` or `SUPERADMIN`). Role priority is
                `USER < ADMIN < SUPERADMIN`.
        ServiceBearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: JWT whose `type` claim is `service`.
    # ── Parameters ────────────────────────────
    parameters:
        eventId:
            name: eventId
            in: path
            required: true
            description: |
                Event identifier supplied as free text so newly registered events can be used.
                Current examples include `EV_GITHUB_ISSUES_BY_COLUMN`, `EV_GITHUB_PR_MERGED`,
                `EV_ZENHUB_ISSUES_BY_COLUMN`, and `EV_BLUEJAY_LOGS_BY_LEVEL`.
            schema:
                type: string
                example: EV_GITHUB_ISSUES_BY_COLUMN
        expand:
            name: expand
            in: query
            required: false
            description: If true, includes raw fetch data in the response.
            schema:
                type: boolean
                default: false
                example: false
...
    requestBodies:
        ProcessEventBody:
            description: Request body for event processing
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ProcessEventRequest'
        ValidateEventBody:
            description: Request body for event validation (all fields optional)
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ValidateEventRequest'
...
    responses:
        Unauthorized:
            description: Missing, malformed, invalid, expired, or unsupported bearer token.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    examples:
                        missingOrMalformed:
                            summary: Missing or malformed Authorization header
                            value:
                                success: false
                                message: Authorization header missing or malformed
                                httpStatus: 401
                                appCode: UNAUTHORIZED
                                data: null
                                error:
                                    message: Authorization header missing or malformed
                        invalidOrExpired:
                            summary: Invalid or expired token
                            value:
                                success: false
                                message: Invalid or expired token
                                httpStatus: 401
                                appCode: UNAUTHORIZED
                                data: null
                                error:
                                    message: Invalid or expired token
        Forbidden:
            description: Authenticated user does not have the required system role.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Insufficient permissions
                        httpStatus: 403
                        appCode: FORBIDDEN
                        data: null
                        error:
                            message: Insufficient permissions
        ValidationError:
            description: Validation error.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Validation failed
                        httpStatus: 400
                        appCode: VALIDATION_ERROR
                        data: null
                        error:
                            message: Validation failed
                            details:
                                - msg: date is required
                                  path: date
        NotFound:
            description: Requested resource not found.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Event EV_UNKNOWN not found
                        httpStatus: 404
                        appCode: NOT_FOUND
                        data: null
                        error:
                            message: Event EV_UNKNOWN not found
        ExternalServiceError:
            description: The Fetcher service is unavailable or failed while validating or obtaining data.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Fetcher service is not available
                        httpStatus: 422
                        appCode: EXTERNAL_SERVICE_ERROR
                        data: null
                        error:
                            message: Fetcher service is not available
        InternalError:
            description: Internal server error.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Unexpected error
                        httpStatus: 500
                        appCode: UNKNOWN_ERROR
                        data: null
                        error:
                            message: Unexpected error

    # ── Schemas ───────────────────────────────
    schemas:
        JsonValue:
            description: Any JSON-compatible value, including null.
            anyOf:
                - type: object
                  nullable: true
                  additionalProperties: true
                - type: array
                  items: {}
                - type: string
                - type: number
                - type: boolean
...
        SuccessResponse:
            type: object
            additionalProperties: false
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: OK
                httpStatus:
                    type: integer
                    example: 200
                appCode:
                    type: string
                    example: SUCCESS
                data:
                    $ref: '#/components/schemas/JsonValue'
                error:
                    type: object
                    nullable: true
                    example: null
            required:
                - success
                - message
                - httpStatus
                - appCode
                - data
                - error
        ErrorResponse:
            type: object
            additionalProperties: false
            properties:
                success:
                    type: boolean
                    example: false
                message:
                    type: string
                    example: Validation failed
                httpStatus:
                    type: integer
                    example: 400
                appCode:
                    type: string
                    example: VALIDATION_ERROR
                data:
                    type: object
                    nullable: true
                    example: null
                error:
                    type: object
                    properties:
                        message:
                            type: string
                            example: Validation failed
                        details:
                            $ref: '#/components/schemas/JsonValue'
                    required:
                        - message
            required:
                - success
                - message
                - httpStatus
                - appCode
                - data
                - error
        Event:
            type: object
            description: Serialized event definition; the executable `process` function is omitted from JSON.
            additionalProperties: false
            properties:
                id:
                    type: string
                    description: Registered event identifier supplied as free text.
                    example: EV_GITHUB_ISSUES_BY_COLUMN
                moreInfo:
                    type: object
                    additionalProperties: false
                    properties:
                        title:
                            type: string
                            example: GitHub Issues by Column
                        description:
                            type: string
                            example: Number of issues in the specified GitHub project columns.
                        example:
                            type: string
                            example: If the Done column contains 10 issues, the metric value would be 10.
                    required:
                        - title
                        - description
                        - example
                fetcherIds:
                    type: array
                    items:
                        type: string
                    example:
                        - FT_GQL_GITHUB_PROJECTV2_ITEMS
                processConfigSchema:
                    type: object
                    description: Serialized zod schema used to validate processConfig.
                    additionalProperties: true
                processScript:
                    type: string
                    description: Stringified process function for documentation/debugging.
            required:
                - id
                - moreInfo
                - fetcherIds
                - processConfigSchema
                - processScript
...
        FetchSummary:
            type: object
            properties:
                fetcherId:
                    type: string
                    example: FT_GQL_GITHUB_PROJECTV2_ITEMS
                fetcherConfig:
                    type: object
                    additionalProperties: true
                fetchResultId:
                    type: string
                    pattern: '^[a-fA-F0-9]{24}$'
                    example: 67c9ff7b2ddf1b3940f9f321
                status:
                    type: string
                    enum: [IN_PROGRESS, COMPLETED, UNAVAILABLE, FAILED]
                    example: COMPLETED
                unavailableReason:
                    type: string
                    nullable: true
            required:
                - fetcherId
                - fetcherConfig
                - fetchResultId
                - status
                - unavailableReason
        FetchDetail:
            allOf:
                - $ref: '#/components/schemas/FetchSummary'
                - type: object
                  properties:
                      data:
                          $ref: '#/components/schemas/JsonValue'
                  required:
                      - data
        Window:
            type: object
            additionalProperties: false
            properties:
                anchorDate:
                    type: string
                    format: date-time
                    description: ISO 8601 date between 2000-01-01 and 2100-01-01.
                    example: '2026-04-22T10:00:00.000Z'
                period:
                    type: array
                    minItems: 1
                    items:
                        type: object
                        additionalProperties: false
                        properties:
                            unit:
                                type: string
                                enum: [millisecond, second, minute, hour, day, week]
                                example: day
                            value:
                                type: integer
                                minimum: 1
                                example: 30
                        required:
                            - unit
                            - value
            required:
                - anchorDate
                - period
...
        ProcessEventRequest:
            type: object
            additionalProperties: false
            properties:
                temporalContext:
                    $ref: '#/components/schemas/TemporalContext'
                window:
                    $ref: '#/components/schemas/Window'
                fetcherConfigs:
                    type: array
                    minItems: 1
                    items:
                        $ref: '#/components/schemas/FetcherConfig'
                processConfig:
                    type: object
                    additionalProperties: true
                    example:
                        columns: [Done]
            required:
                - temporalContext
                - window
                - fetcherConfigs
                - processConfig
        ValidateEventRequest:
            type: object
            description: Both `fetcherConfigs` and `processConfig` are optional.
            additionalProperties: false
            properties:
                fetcherConfigs:
                    type: array
                    items:
                        $ref: '#/components/schemas/FetcherConfig'
                processConfig:
                    type: object
                    additionalProperties: true
                    example:
                        columns: [Done]
...
        ProcessedEventSummary:
            type: object
            additionalProperties: false
            properties:
                events:
                    type: array
                    nullable: true
                    items:
                        type: object
                        additionalProperties: true
                eventId:
                    type: string
                    example: EV_GITHUB_ISSUES_BY_COLUMN
                date:
                    type: string
                    format: date-time
                    example: '2026-04-22T10:00:00.000Z'
                window:
                    $ref: '#/components/schemas/Window'
                fetchs:
                    type: array
                    items:
                        $ref: '#/components/schemas/FetchSummary'
                processConfig:
                    type: object
                    additionalProperties: true
            required:
                - events
                - eventId
                - date
                - window
                - fetchs
                - processConfig
        ProcessedEventDetailed:
            type: object
            additionalProperties: false
            properties:
                events:
                    type: array
                    nullable: true
                    items:
                        type: object
                        additionalProperties: true
                eventId:
                    type: string
                    example: EV_GITHUB_ISSUES_BY_COLUMN
                date:
                    type: string
                    format: date-time
                    example: '2026-04-22T10:00:00.000Z'
                window:
                    $ref: '#/components/schemas/Window'
                fetchs:
                    type: array
                    items:
                        $ref: '#/components/schemas/FetchDetail'
                processConfig:
                    type: object
                    additionalProperties: true
            required:
                - events
                - eventId
                - date
                - window
                - fetchs
                - processConfig
        EventValidationIssue:
            type: object
            description: Validation issue from fetcherConfig or processConfig.
            additionalProperties: true
            properties:
                source:
                    type: string
                    enum: [fetcherConfig, processConfig]
                    example: processConfig
                fetcherId:
                    type: string
                    example: FT_GQL_GITHUB_PROJECTV2_ITEMS
                code:
                    type: string
                    example: invalid_type
                message:
                    type: string
                    example: Required
                path:
                    type: array
                    items:
                        oneOf:
                            - type: string
                            - type: integer
            required:
                - source
                - code
                - message
                - path
        EventValidationResult:
            oneOf:
                - type: object
                  properties:
                      valid:
                          type: boolean
                          enum: [true]
                          example: true
                  required:
                      - valid
                - type: object
                  properties:
                      valid:
                          type: boolean
                          enum: [false]
                          example: false
                      error:
                          type: string
                          example: Invalid processConfig
                      issues:
                          type: array
                          items:
                              $ref: '#/components/schemas/EventValidationIssue'
                  required:
                      - valid
                      - error
...